External CA (keyless) verification.
Allow keyless verification when signing with a public-CA backed cert.

Use case - key on HSM, cert from well-known CA (e.g. DigiCert)
Cosign signing with PKCS#11 URL for key - and certs/chain.

There are *no* SANS or Extensions->issuer in this case - so keyless
fails. To make it work a bit, we consider the 'subject' when there are
no sans, and the main issuer, when there are no additional issuer info
in x.509 cert extensions.